Condense storage path handling
