+++ app/src/main/feature/stores/common/StoreInstallPathSafety.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.net.Uri
 import android.os.Environment
 import com.winlator.cmod.feature.stores.steam.utils.PrefManager
+import com.winlator.cmod.shared.android.StoragePathUtils
 import com.winlator.cmod.shared.io.FileUtils
 import java.io.File
 
@@ -22,7 +23,7 @@ object StoreInstallPathSafety {
             return DeleteCheck(false, "empty install path")
         }
 
-        val target = normalize(File(targetPath))
+        val target = StoragePathUtils.normalize(File(targetPath)) ?: File(targetPath).absoluteFile
         val protected = buildProtectedRoots(context, protectedRoots)
         val matchedRoot = protected.firstOrNull { samePath(target, it) }
         if (matchedRoot != null) {
@@ -45,7 +46,7 @@ object StoreInstallPathSafety {
         val roots = linkedMapOf<String, File>()
 
         fun add(file: File?) {
-            val normalized = file?.let(::normalize) ?: return
+            val normalized = StoragePathUtils.normalize(file) ?: return
             if (normalized.path.isBlank()) return
             roots[normalized.path] = normalized
         }
@@ -55,22 +56,12 @@ object StoreInstallPathSafety {
         add(File("/mnt/media_rw"))
         add(Environment.getExternalStorageDirectory())
 
-        File("/storage").listFiles().orEmpty().forEach { child ->
-            if (!child.isDirectory || child.name == "self") return@forEach
-            if (child.name == "emulated") {
-                add(File(child, "0"))
-            } else {
-                add(child)
-            }
-        }
-
-        File("/mnt/media_rw").listFiles().orEmpty().forEach { child ->
-            if (child.isDirectory) add(child)
-        }
-
-        context?.getExternalFilesDirs(null).orEmpty().forEach { externalFilesDir ->
-            add(resolveStorageRootFromExternalFilesDir(externalFilesDir))
-        }
+        StoragePathUtils.getMountedStorageRoots(
+            context = context,
+            includePrimary = true,
+            includeMediaRw = true,
+            requireBrowsable = false,
+        ).forEach(::add)
 
         configuredDownloadRoots(context).forEach(::add)
         extraRoots.map(::File).forEach(::add)
@@ -115,23 +106,8 @@ object StoreInstallPathSafety {
         return File(uriPath ?: value)
     }
 
-    private fun resolveStorageRootFromExternalFilesDir(dir: File): File? {
-        val absolute = dir.absoluteFile
-        val androidDir =
-            generateSequence(absolute) { it.parentFile }
-                .firstOrNull { it.name.equals("Android", ignoreCase = true) }
-        return androidDir?.parentFile
-    }
-
-    private fun normalize(file: File): File =
-        try {
-            file.canonicalFile
-        } catch (_: Exception) {
-            file.absoluteFile
-        }
-
     private fun samePath(
         a: File,
         b: File,
-    ): Boolean = normalize(a).path.trimEnd('/') == normalize(b).path.trimEnd('/')
+    ): Boolean = StoragePathUtils.samePath(a, b)
 }
