+++ app/src/main/shared/android/DirectoryPickerDialog.kt
@@ -12,8 +12,6 @@ import android.view.ViewGroup
 import android.view.Window
 import android.view.WindowManager
 import android.widget.Toast
-import android.os.storage.StorageManager
-import android.os.storage.StorageVolume
 import androidx.compose.animation.core.CubicBezierEasing
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
@@ -263,7 +261,7 @@ object DirectoryPickerDialog {
 
     private fun normalizeAllowedExtensions(allowedExtensions: Set<String>): Set<String> =
         allowedExtensions
-            .map { it.trim().trimStart('.').lowercase(Locale.getDefault()) }
+            .map { it.trim().trimStart('.').lowercase(Locale.ROOT) }
             .filter { it.isNotEmpty() }
             .toSet()
 
@@ -869,43 +867,33 @@ object DirectoryPickerDialog {
             }
 
         val children =
-            listSortedChildren(currentDir) { child ->
-                child.isDirectory && canBrowse(child)
-            }
+            currentDir
+                .listFiles()
+                .orEmpty()
+                .asSequence()
+                .sortedWith(compareBy<File>({ it.isHidden }, { it.name.lowercase(Locale.ROOT) }))
+                .toList()
 
-        for (child in children) {
-            entries += Entry(label = entryLabel(child), target = child)
-        }
+        children
+            .filter { it.isDirectory && canBrowse(it) }
+            .forEach { child ->
+                entries += Entry(label = entryLabel(child), target = child)
+            }
 
         if (mode == SelectionMode.FILE) {
-            val files =
-                listSortedChildren(currentDir) { child ->
-                    child.isFile && canSelectFile(child, allowedExtensions)
+            children
+                .filter { it.isFile && canSelectFile(it, allowedExtensions) }
+                .forEach { file ->
+                    entries += Entry(
+                        label = entryLabel(file),
+                        target = file,
+                        isSelectableFile = true,
+                    )
                 }
-
-            for (file in files) {
-                entries += Entry(
-                    label = entryLabel(file),
-                    target = file,
-                    isSelectableFile = true,
-                )
-            }
         }
 
         return entries
     }
-
-    private fun listSortedChildren(
-        currentDir: File,
-        predicate: (File) -> Boolean,
-    ): List<File> =
-        currentDir
-            .listFiles()
-            ?.asSequence()
-            ?.filter(predicate)
-            ?.sortedWith(compareBy<File>({ it.isHidden }, { it.name.lowercase(Locale.getDefault()) }))
-            ?.toList()
-            ?: emptyList()
 
     private fun entryLabel(file: File): String = file.name.ifBlank { file.absolutePath }
 
@@ -924,181 +912,17 @@ object DirectoryPickerDialog {
         return roots.firstOrNull() ?: Environment.getExternalStorageDirectory()
     }
 
-    private fun buildRootDirectories(activity: Activity): List<File> {
-        val roots = linkedMapOf<String, File>()
-        val storageManager = activity.getSystemService(StorageManager::class.java)
-        val externalFilesDirs =
-            activity.getExternalFilesDirs(null)
-                .filterNotNull()
-                .filter { isReadableMountedState(Environment.getExternalStorageState(it)) }
-
-        fun addRoot(dir: File?) {
-            val normalized = dir?.absoluteFile ?: return
-            if (!canBrowse(normalized)) return
-            roots.putIfAbsent(normalized.absolutePath, normalized)
-        }
-
-        fun addResolvedRoot(
-            root: File?,
-            browseSeed: File? = null,
-        ) {
-            resolveBrowsableRoot(root, browseSeed)?.let(::addRoot)
-        }
-
-        addResolvedRoot(Environment.getExternalStorageDirectory())
-
-        storageManager?.storageVolumes.orEmpty()
-            .filter(::isReadableVolume)
-            .forEach { volume ->
-                val volumeExternalDirs =
-                    externalFilesDirs.filter { dir ->
-                        belongsToVolume(storageManager, dir, volume)
-                    }
-
-                val volumeCandidates =
-                    buildList {
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-                            add(volume.directory)
-                        }
-                        volume.uuid?.let { uuid ->
-                            add(File("/storage/$uuid"))
-                            add(File("/mnt/media_rw/$uuid"))
-                        }
-                        volumeExternalDirs.forEach { dir ->
-                            add(resolveStorageRootFromExternalFilesDir(dir))
-                        }
-                    }
-
-                var resolvedRoot: File? = null
-                for (candidate in volumeCandidates) {
-                    resolvedRoot =
-                        volumeExternalDirs
-                            .asSequence()
-                            .mapNotNull { dir -> resolveBrowsableRoot(candidate, dir) }
-                            .firstOrNull()
-                            ?: resolveBrowsableRoot(candidate)
-                    if (resolvedRoot != null) {
-                        break
-                    }
-                }
-
-                if (resolvedRoot == null) {
-                    resolvedRoot =
-                        volumeExternalDirs
-                            .asSequence()
-                            .mapNotNull(::resolveFallbackBrowsableRootFromExternalFilesDir)
-                            .firstOrNull()
-                }
-
-                addRoot(resolvedRoot)
-            }
-
-        File("/storage").listFiles().orEmpty().forEach { child ->
-            if (!child.isDirectory || child.name == "self") return@forEach
-            if (child.name == "emulated") {
-                addResolvedRoot(File(child, "0"))
-            } else {
-                addResolvedRoot(child)
-            }
-        }
-
-        externalFilesDirs
-            .mapNotNull(::resolveStorageRootFromExternalFilesDir)
-            .forEach { root ->
-                addResolvedRoot(root)
-            }
-
-        externalFilesDirs
-            .mapNotNull(::resolveFallbackBrowsableRootFromExternalFilesDir)
-            .forEach(::addRoot)
-
-        return roots.values.toList()
-    }
-
-    private fun resolveStorageRootFromExternalFilesDir(dir: File): File? {
-        val absolute = dir.absoluteFile
-        val androidDir =
-            generateSequence(absolute) { it.parentFile }
-                .firstOrNull { it.name.equals("Android", ignoreCase = true) }
-        if (androidDir?.parentFile != null) {
-            return androidDir.parentFile
-        }
-
-        return generateSequence(absolute) { it.parentFile }
-            .drop(4)
-            .firstOrNull()
-    }
-
-    private fun resolveBrowsableRoot(
-        root: File?,
-        browseSeed: File? = null,
-    ): File? {
-        val normalizedRoot = root?.absoluteFile ?: return null
-        if (canBrowse(normalizedRoot)) return normalizedRoot
-        return highestBrowsablePathWithinRoot(normalizedRoot, browseSeed)
-    }
-
-    private fun resolveFallbackBrowsableRootFromExternalFilesDir(dir: File): File? {
-        val absolute = dir.absoluteFile
-        val resolvedRoot = resolveStorageRootFromExternalFilesDir(absolute)
-        return when {
-            resolvedRoot != null -> highestBrowsablePathWithinRoot(resolvedRoot, absolute)
-            canBrowse(absolute) -> absolute
-            else ->
-                generateSequence(absolute.parentFile) { it.parentFile }
-                    .firstOrNull(::canBrowse)
-        }
-    }
-
-    private fun highestBrowsablePathWithinRoot(
-        root: File,
-        browseSeed: File?,
-    ): File? {
-        val normalizedRoot = root.absoluteFile
-        val normalizedSeed = browseSeed?.absoluteFile ?: return null
-        var current: File? = normalizedSeed
-        var best: File? = null
-        while (current != null && isSameOrDescendant(current, normalizedRoot)) {
-            if (canBrowse(current)) {
-                best = current
-            }
-            current = current.parentFile
-        }
-        return best
-    }
-
-    private fun belongsToVolume(
-        storageManager: StorageManager,
-        dir: File,
-        volume: StorageVolume,
-    ): Boolean {
-        val dirVolume = storageManager.getStorageVolume(dir) ?: return false
-        if (dirVolume.isPrimary != volume.isPrimary) return false
-
-        val dirUuid = dirVolume.uuid
-        val volumeUuid = volume.uuid
-        return if (!dirUuid.isNullOrBlank() || !volumeUuid.isNullOrBlank()) {
-            dirUuid == volumeUuid
-        } else {
-            true
-        }
-    }
-
-    private fun isReadableVolume(volume: StorageVolume): Boolean = isReadableMountedState(volume.state)
-
-    private fun isReadableMountedState(state: String?): Boolean =
-        state == Environment.MEDIA_MOUNTED || state == Environment.MEDIA_MOUNTED_READ_ONLY
+    private fun buildRootDirectories(activity: Activity): List<File> =
+        StoragePathUtils.buildBrowsableStorageRoots(activity)
 
     private fun isSameOrDescendant(
         candidate: File,
         root: File,
     ): Boolean {
-        val candidatePath = candidate.absolutePath.trimEnd('/')
-        val rootPath = root.absolutePath.trimEnd('/')
-        return candidatePath == rootPath || candidatePath.startsWith("$rootPath/")
+        return StoragePathUtils.isSameOrDescendant(candidate, root)
     }
 
-    private fun canBrowse(dir: File?): Boolean = dir != null && dir.exists() && dir.isDirectory && dir.canRead()
+    private fun canBrowse(dir: File?): Boolean = StoragePathUtils.canBrowse(dir)
 
     private fun canSelectFile(
         file: File,
@@ -1106,7 +930,7 @@ object DirectoryPickerDialog {
     ): Boolean {
         if (!file.canRead()) return false
         if (allowedExtensions.isEmpty()) return true
-        return allowedExtensions.contains(file.extension.lowercase(Locale.getDefault()))
+        return allowedExtensions.contains(file.extension.lowercase(Locale.ROOT))
     }
 
     private fun ensureAllFilesAccess(activity: Activity): Boolean {
